disp: msm: dp: add force bond mode support to different monitors
Bond mode requires mode and link parameters to be exactly the
same and primary monitor's prameters will be used for other bonded
displays.

For standard dual-dp displays this is OK as their parameters are
always the same. However for force mode, different monitors might
be connected and cause unexpected behavior.

This change added more checks to get the common parameters from
different monitors for force mode and relaxed the requirement.

Change-Id: Ib8e2323dc5958db52e59aa8483f257c0fca4a247
Signed-off-by: Xiaowen Wu <wxiaowen@codeaurora.org>
Signed-off-by: Jim Wang <quic_jimwan@quicinc.com>