disp: msm: misr: misr fence implementation and timeline support
Based on the post-commit fence framework to implement misr fence
for polling misr state and reading signature from user space.

Timeline is supported to avoid that fence is not signaled in
some extreme cases. For example, the interrupt is not triggered
due to a hardware error or software problem, even in this case,
all related misr fences should be signaled. That is to say, any
fence object will be signaled within a reasonable period should
be guaranteed, and it must be ensured that the returned signature
is matched with each atomic commit.

Change-Id: Ibf8f29e4dcdc64c1091048f472332285f7e4ff9d
Signed-off-by: Zhiqiang Liu <zhiqliu@codeaurora.org>
Signed-off-by: Zhiqiang Liu <quic_zhiqliu@quicinc.com>