disp: msm-hyp: sort enumeration ports by port id
If there are two DPUs on some platforms, We always enumerate all the
ports on device0 firstly, followed by the ports on device1 even if
we configure the ports on device1 firstly in display xml, which is
not aligned with some APPs.

Therefor, sort connector list by port id after enumerating all the
ports on both devices.

Change-Id: Ib253d8abe1f3d1f68f851622e1ef6a34f7ab38d8
Signed-off-by: Shilun Wan <quic_shilwan@quicinc.com>