disp: msm: fence: fix generic fence list is empty issue
The sde_fence_helper_signal is designed for quick response,
but atomicity cannot be guaranteed in this case.

In the vsync interrupt case, since vsync comes periodically,
the bottom-half of interrupts are called sequentially so it
works fine.

But for misr case, it will use mismatch interrupt and vsync
interrupt to quickly respond the MISR events. The misr irqs
can come at any time and the scheduling of workqueue depends
on the scheduler.

So below call-sequence will trigger this issue,
A indicates commit thread, B indicates workqueue,
1. [A] prepare misr/post-commit fence
2. [B] move all generic fences to local fence list
3. [A] find misr fence and do update
4. [B] move back the un-signaled fences to generic fence list

In above third step, it cannot find the available fence, so
failed to do update. The fixed method is that remove the list
move operation and make the fence list operation is atomic.

Change-Id: I54fb0b406a36406fa1535fe3139ea141c7fb2dd3
Signed-off-by: Zhiqiang Liu <quic_zhiqliu@quicinc.com>