disp: msm: dp: skip pm if DP is not fully initialized
If a DP display is defined in device tree, but not included in the
connectors list in device tree, the DP driver will not be fully
initialized, including its sub modules.

In this case, PM suspend/complete/freeze should be skiped to avoid
invalid pointer access. By checking the base_connector validity could
know if the DP driver is fully initialized.

Change-Id: Ibd04a02470688a33eee4f0ec335e2942332230e2
Signed-off-by: Jim Wang <quic_jimwan@quicinc.com>