disp: msm: dp: fix "exceeds max dp count" issue
DP driver will fail to call component_add in some cases, the global
dp_display variables should be set to NULL because the dp_display
instance needs to be released and it will be reallocated in next probe
time. If not, all global dp_display slots will be occupied illegally
and this will cause dp driver probe failure issue.

Change-Id: Ia0d9e50c294156af47c46cab1fc330a717c5dcb7
Signed-off-by: Zhiqiang Liu <quic_zhiqliu@quicinc.com>