disp: msm: dp: fix an issue AUX sim generated gargage EDID
Before the EDID is ready, the AUX simulator will not be able to copy valid
EDID to the target memory when SDE request the EDID, which end up with
corrupted EDID in the memory and lead to code crash in connector_bad_edid.

Change the code fill in zeros to the target memory when the request EDID
addr/size is out of range.

Change-Id: I89fd8f506dd25052c5ac8de335e0884a92244227
Signed-off-by: Jim Wang <quic_jimwan@quicinc.com>