msm/dp: dp_mst_sim: fix error during debugfs init
When CONFIG_DEBUG_FS is disabled the call to create
the debugfs would fail. debugfs_create_dir will
result in an error which leads to failure in the
DP Initialization.

The change adds a check to read if the config is
defined and skips the error and prevents the
failure of DP Initialization.

Change-Id: Ide7a1d95dca611de9aa61d7d2d72f2cbf0215787
Signed-off-by: Rahul Sharma <quic_rahsha@quicinc.com>