disp: msm: avoid using 'timespec'
The timespec structure and associated interfaces are deprecated and will
be removed in the future because of the y2038 overflow.
The use of ktime_to_timespec() in timeout_to_jiffies() does not
suffer from that overflow, but is easy to avoid by just converting
the ktime_t into jiffies directly.

Change-Id: Ic27b53791e68ad516fa190a3e7f3b5824ad3635c
Reviewed-by: Jordan Crouse <jcrouse@codeaurora.org>
Signed-off-by: Arnd Bergmann <arnd@arndb.de>
Git-commit: e0b41e5921c553383ae9d241c41bd8f9b5e06089
Git-repo: git://git.kernel.org/pub/scm/linux/kernel/git/arnd/playground.git
Signed-off-by: Parikshit Pareek <quic_ppareek@quicinc.com>
Signed-off-by: ruoyyang <quic_ruoyyang@quicinc.com>