disp: msm: dp: for DP MST read EDID earlier
For DP MST, the EDID is read when get_modes is called. But for bonded DP MST,
it will be too late since the sibling connector EDID is not retrieved when
the get_modes is called to the master connector. This will result no sibling
connector is found for the first time get_modes is called, but can be found
for the reset of time, meaning inconsistent result output of get_mdoes.

Move the EDID reading to the connector register callback, which is called
after LINK_ADDRESS and ENUM_PATH_RESOURCES.

Change-Id: Ica7eee80a64488e2c248e09ce064040186d6d90b
Signed-off-by: Jim Wang <quic_jimwan@quicinc.com>