disp: msm: sde: get the correct display pointer for SHD
This change is to get the correct display pointer in case of SHD.
If the connector is shared, the display should be a shared.
The dsi_base of the SHD should be accessed by the sde_connector APIs
instead of typecasting the shd_display data structure directly
as dsi_display, which is different format and size.
Otherwise it will cause out of range memory access issue.

change-Id: I85f4ccc457cb4e5e3c7a3e0cc5df8a162da5ecf0
Signed-off-by: Ramya SR <quic_rsr@quicinc.com>
Signed-off-by: Rahul Sharma <quic_rahsha@quicinc.com>