disp: msm: sde: fix issue resource not release correctly
For test only commit, resource reserved is commit and pending flag is
marked false, then the releasing for the test only will not find the
pending reservation, so the releasing is skipped, lead to resource
leaking. When other reservation request comes, there will be no enough
hardware resource available.

For test only commit, leave the pending flag to true solve the issue.

Change-Id: Id591c90ac1fa04d3e2a2d292087f4071fe525d45
Signed-off-by: Jim Wang <quic_jimwan@quicinc.com>